feat(Util): Add getUserIDFromMention and fetchUserFromMention
The getUserFromMention is flawed, and can lead to ratelimits on fetchUser endpoint.
The ratelimit is caused by everyone mentioning someone and the bot fetches that mentioned user.
The truth is: fetching the User from mention is unnecessary.
Because it is used to extract the ID of user in mention in every message and check if it's Bot's User ID